chatflow/requirements.txt
openai
jinja2
python-dotenv
tiktoken
promptflow[azure]
promptflow-tools